Add SQLite support (#9)
Allow Pecorino throttles and leaky buckets to be defined using SQLite. Note that it is not the SQLite "time" value that we use, but the Ruby value - it has a finer resolution and is easier to work with, and should be harmless since SQLite is in-process anyway.

The same could potentially be done with SQLite date and time functions, but I found them very hard to work with for getting time intervals - and the best resolution I could obtain was milliseconds, whereas Pecorino benefits from time resolution being finer than that.

The change moves the DB statements used into separate modules, one per DB. Since we only got 2 now I opted for inheritance in tests - to avoid duplication. Once we get a third one in we might switch to using a module.

Closes #4