Extract Adapters, add Redis adapter (#18)
This also adds a Redis adapter and a Memory adapter. The Redis adapter is useful since it can be a continuation of Prorate in a way - but with conditional fillup added in.

The memory store is useful in a different way - it can be used as a reference implementation to test other adapters against. Also make the adapter configurable on the `Pecorino` module.