Hide the AR connection behind a shim
ActiveRecord connection management can be tricky, and we don't need it as such - we need to _integrate_ with it well.

Also, all of our queries are .uncached. We do still want the queries to go through ActiveRecord adapters, because there is useful instrumentation in their `execute` which some folks are using

Relocate sanitization into the shim as well, for future work